Whole lotta changes

+++ contracts/interfaces/ISmolsTraitStorage.sol
@@ -4,24 +4,24 @@ pragma solidity ^0.8.0;
 import "../libraries/SmolsLibrary.sol";
 
 interface ISmolsTraitStorage {
-    function traits(uint256 _traitId, uint8 _dependencyLevel) external view returns(Trait memory);
+    function traits(uint256 _traitId, uint256 _dependencyLevel) external view returns(Trait memory);
 
-    function getTrait(uint256 _traitId, uint8 _dependencyLevel)
+    function getTrait(uint256 _traitId, uint256 _dependencyLevel)
         external
         view
         returns (Trait memory);
 
-    function getTraitType(uint256 _traitId, uint8 _dependencyLevel)
+    function getTraitType(uint256 _traitId, uint256 _dependencyLevel)
         external
         view
         returns (bytes memory);
         
-    function getTraitName(uint256 _traitId, uint8 _dependencyLevel)
+    function getTraitName(uint256 _traitId, uint256 _dependencyLevel)
         external
         view
         returns (bytes memory);
 
-    function getTraitImage(uint256 _traitId, uint8 _dependencyLevel)
+    function getTraitImage(uint256 _traitId, uint256 _dependencyLevel)
         external
         view
         returns (bytes memory);
